+++ pocketutils/tools/program_tools.py
@@ -3,7 +3,7 @@ import subprocess
 from dataclasses import dataclass
 
 from pocketutils.core import PathLike
-from pocketutils.core.exceptions import CalledProcessError, ParsingError
+from pocketutils.core.exceptions import ParsingError
 from pocketutils.tools.base_tools import BaseTools